requirements.txt
absl-py==2.1.0
accelerate==1.13.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.0
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
Authlib==1.5.2
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
docker-pycreds==0.4.0
exceptiongroup==1.2.2
executing==2.0.1
extra-streamlit-components==0.1.71
fastapi==0.115.3
fastjsonschema==2.20.0
ffmpy==0.4.0
filelock==3.15.4
Flask==3.1.0
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.6.1
fpdf2
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-api-python-client==2.141.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
googleapis-common-protos==1.63.2
gradio==5.20.0
gradio_modal
gradio_client==1.7.2
greenlet==3.1.1
groovy==0.1.2
grpcio==1.65.1
h11==0.14.0
h5py==3.16.0
hf-xet==1.4.2
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
httpx-sse==0.4.0
huggingface_hub==1.7.2
idna==3.7
ImageIO==2.37.3
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.13.0
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterhub-dummyauthenticator==0.3.1
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
keras==3.4.1
kiwisolver==1.4.5
langchain-openai
langchain-text-splitters
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith
latex2mathml==3.77.0
lazy_loader==0.4
libclang==18.1.1
linkify-it-py==2.0.3
Markdown==3.6
markdown-it-py==2.2.0
markdown2==2.5.1
MarkupSafe==2.1.5
marshmallow==3.23.1
matplotlib==3.10.8
matplotlib-inline==0.1.7
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
namex==0.0.8
narwhals==2.18.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvcc-cu12==12.3.107
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
openai==2.29.0
opt-einsum==3.3.0
optree==0.12.1
orjson==3.11.7
ormsgpack==1.12.2
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
pip-tools==7.4.1
platformdirs==4.2.2
plotly==6.6.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
propcache==0.2.0
proto-plus==1.24.0
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.2
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.20
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.1.0
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
regex==2026.2.28
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
ruff==0.11.0
safehttpx==0.1.6
safetensors==0.4.3
scikit-image==0.25.2
scikit-learn==1.5.1
scipy==1.14.0
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==2.17.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.5
SQLAlchemy==2.0.35
stack-data==0.6.3
starlette==0.41.0
streamlit==1.37.1
streamlit-google-auth==1.1.8
svgwrite==1.4.3
sympy==1.13.1
tenacity==8.5.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2025.3.13
tiktoken==0.12.0
tinycss2==1.3.0
tokenizers==0.22.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.13.2
torch==2.6.0
torchvision==0.21.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
triton==3.2.0
typer==0.15.2
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.15.0
tzdata==2024.1
uc-micro-py==1.0.3
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
uuid_utils==0.14.1
uvicorn==0.32.0
wandb==0.18.5
watchdog==4.0.2
wavedrom==2.0.3.post3
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.1.3
widgetsnbextension==4.0.11
wrapt==1.16.0
xxhash==3.6.0
yarl==1.16.0
zstandard==0.25.0
